lms-backend-main/requirements.txt
asgiref==3.3.4
autopep8==2.0.1
boto3==1.26.38
botocore==1.29.38
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
Django==3.2.16
django-cors-headers==3.13.0
django-crispy-forms==1.11.2
django-filter==22.1
django-tenant-schemas==1.11.0
django-tenants==3.4.5
django-zoom-meetings==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
ordered-set==4.1.0
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.1
tqdm==4.64.1
urllib3==1.26.12
